fix(sensitivity): natural-units SensResult multipliers; dedupe dense_to_vec (F1/F8 nits)
SensResult.mult_g/mult_x_l/mult_x_u were captured via pack_*_for_user,
which does not divide by obj_scale_factor — so when NLP scaling fired
they were obj-scaled, and the doc claiming 'same convention as the C
ABI / Python info dict' was stale after the F1 finalize fix. Capture
now uses the finalize_solution_lambda/_z_l/_z_u trait methods (natural
units), matching every other user-facing multiplier surface and the
warm-start round trip (initialize_starting_point multiplies back by
obj_scale on the way in). Consumers audited: pounce-py's
solve_with_sens reads the bridge finalize, not these fields;
diff_handoff's documented convention now actually holds. New test
captured_multipliers_are_invariant_to_nlp_scaling asserts mult_g under
gradient-based scaling equals both the unscaled run and the analytic
value (pre-fix it was off by the ~1/1200 obj scale).

Also dedupe the two identical dense_to_vec copies (solver.rs,
convenience.rs) into a shared crate-private vec_util module so the
single implementation carries the compound-flatten tests — the F8
verification flagged the convenience.rs copy as untested.
